+++ src/main/java/com/buuz135/jeivillagers/jei/VillagerPlugin.java
@@ -2,10 +2,13 @@ package com.buuz135.jeivillagers.jei;
 
 import com.buuz135.jeivillagers.FakeVillager;
 import com.buuz135.jeivillagers.Jeivillagers;
-import mezz.jei.api.IJeiRuntime;
-import mezz.jei.api.IModPlugin;
-import mezz.jei.api.IModRegistry;
-import mezz.jei.api.JEIPlugin;
+import com.buuz135.jeivillagers.jei.ingredient.VillageCareerRenderer;
+import com.buuz135.jeivillagers.jei.ingredient.VillagerCareerFactory;
+import com.buuz135.jeivillagers.jei.ingredient.VillagerCareerHelper;
+import com.google.common.base.Preconditions;
+import mezz.jei.api.*;
+import mezz.jei.api.ingredients.IModIngredientRegistration;
+import mezz.jei.api.recipe.IFocus;
 import mezz.jei.api.recipe.IRecipeCategoryRegistration;
 import net.minecraft.client.gui.GuiMerchant;
 import net.minecraft.init.Items;
@@ -24,6 +27,20 @@ public class VillagerPlugin implements IModPlugin {
 
 
     private VillagerCategory category;
+    private ISubtypeRegistry subtypes;
+    private static IRecipesGui recipesGui;
+    private static IRecipeRegistry recipeRegistry;
+
+    @Override
+    public void registerItemSubtypes(ISubtypeRegistry subtypeRegistry) {
+        this.subtypes = subtypeRegistry;
+    }
+
+    @Override
+    public void registerIngredients(IModIngredientRegistration registry) {
+        Preconditions.checkState(this.subtypes != null);
+        registry.register(VillagerRegistry.VillagerCareer.class, VillagerCareerFactory.create(), new VillagerCareerHelper(), new VillageCareerRenderer());
+    }
 
     @Override
     public void registerCategories(IRecipeCategoryRegistration registry) {
@@ -33,15 +50,9 @@ public class VillagerPlugin implements IModPlugin {
 
     @Override
     public void register(IModRegistry registry) {
-        NBTTagCompound compound = new NBTTagCompound();
-        NBTTagCompound idnbt = new NBTTagCompound();
-        idnbt.setString("id", "minecraft:villager");
-        compound.setTag("EntityTag", idnbt);
-        ItemStack stack = new ItemStack(Items.SPAWN_EGG);
-        stack.setTagCompound(compound);
-        registry.addRecipeCatalyst(stack, category.getUid());
 
-        long time = System.currentTimeMillis();
+        registry.addRecipeCatalyst(getVillagerEgg(), category.getUid());
+
         tradeInfoMultimap.clear();
         FakeVillager merchant = new FakeVillager();
         Random random = new Random();
@@ -57,7 +68,6 @@ public class VillagerPlugin implements IModPlugin {
         for (Jeivillagers.VillagerTradeInfo tradeInfo : tradeInfoMultimap.values()) {
             tradeInfo.clean();
         }
-        System.out.println("Checked villager recipes in " + (System.currentTimeMillis() - time) + "ms.");
 
         for (VillagerRegistry.VillagerCareer career : tradeInfoMultimap.keySet()) {
             registry.addRecipes(tradeInfoMultimap.get(career).stream().filter(villagerTradeInfo -> !villagerTradeInfo.outputStack.isEmpty()).map(villagerTradeInfo -> new VillagerRecipe(career, villagerTradeInfo)).collect(Collectors.toList()), category.getUid());
@@ -67,6 +77,22 @@ public class VillagerPlugin implements IModPlugin {
 
     @Override
     public void onRuntimeAvailable(IJeiRuntime jeiRuntime) {
+        recipesGui = jeiRuntime.getRecipesGui();
+        recipeRegistry = jeiRuntime.getRecipeRegistry();
+    }
 
+    public static ItemStack getVillagerEgg(){
+        NBTTagCompound compound = new NBTTagCompound();
+        NBTTagCompound idnbt = new NBTTagCompound();
+        idnbt.setString("id", "minecraft:villager");
+        compound.setTag("EntityTag", idnbt);
+        ItemStack stack = new ItemStack(Items.SPAWN_EGG);
+        stack.setTagCompound(compound);
+        return stack;
+    }
+
+    public static void showUses(VillagerRegistry.VillagerCareer career) {
+        if (recipesGui != null && recipeRegistry != null)
+            recipesGui.show(recipeRegistry.createFocus(IFocus.Mode.OUTPUT, career));
     }
 }