+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # 1.0.2
 + Added a clickable area in the villager GUI to show the recipes.
++ All the careers are shown in the JEI Item collection in the right as spawn eggs (The don't spawn any entity).
++ Added a clickable are in the JEI recipe to show all the recipes of one career. (Click in the text: `Career: X`)